refactor(naming): remove cloud from prefix as the plugin now works for both cloud or self-hosted
