--- src/main/java/io/kestra/plugin/prefect/cloud/package-info.java
@@ -1,9 +1,0 @@
-@PluginSubGroup(
-    title = "Prefect Cloud",
-    description = "This sub-group of plugins contains tasks for interacting with Prefect Cloud.",
-    categories = { PluginSubGroup.PluginCategory.TOOL, PluginSubGroup.PluginCategory.CLOUD }
-)
-package io.kestra.plugin.prefect.cloud;
-
-import io.kestra.core.models.annotations.PluginSubGroup;
-
