+++ src/main/java/io/kestra/plugin/prefect/package-info.java
@@ -1,0 +1,9 @@
+@PluginSubGroup(
+    title = "Prefect",
+    description = "This sub-group of plugins contains tasks for interacting with Prefect (both Cloud and self-hosted instances).",
+    categories = { PluginSubGroup.PluginCategory.TOOL }
+)
+package io.kestra.plugin.prefect;
+
+import io.kestra.core.models.annotations.PluginSubGroup;
+
