+++ src/test/java/io/kestra/plugin/prefect/CreateFlowRunTest.java
@@ -1,4 +1,4 @@
-package io.kestra.plugin.prefect.cloud;
+package io.kestra.plugin.prefect;
 
 import io.kestra.core.junit.annotations.KestraTest;
 import io.kestra.core.models.property.Property;
